+++ src/com/jsheets/expressions/Add.java
@@ -1,4 +1,4 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public class Add<T> extends BinaryExpression<T, Number> {
   public Add(Expression<T, Number> left, Expression<T, Number> right) {
