+++ src/com/jsheets/expressions/And.java
@@ -1,4 +1,4 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public class And<T> extends BinaryExpression<T, Boolean> {
   public And(Expression<T, Boolean> left, Expression<T, Boolean> right) {
