+++ src/com/jsheets/expressions/BooleanConstant.java
@@ -1,4 +1,4 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public class BooleanConstant extends ConstantExpression<Boolean> {
   public BooleanConstant(Boolean value) {
