+++ src/com/jsheets/expressions/BooleanExpression.java
@@ -1,3 +1,3 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public interface BooleanExpression extends Expression<Boolean, Boolean> {}
