+++ src/com/jsheets/expressions/ComparisonExpression.java
@@ -1,6 +1,6 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
-public abstract class BinaryExpression<T, R> implements Expression<T, R> {
+public abstract class ComparisonExpression<T, R> implements BooleanExpression {
   private final Expression<T, R> left;
   private final Expression<T, R> right;
 
@@ -12,7 +12,10 @@ public abstract class BinaryExpression<T, R> implements Expression<T, R> {
     return right.compute();
   }
 
-  public BinaryExpression(Expression<T, R> left, Expression<T, R> right) {
+  public ComparisonExpression(
+    Expression<T, R> left,
+    Expression<T, R> right
+  ) {
     this.left = left;
     this.right = right;
   }
