+++ src/com/jsheets/expressions/Divide.java
@@ -1,4 +1,4 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public class Divide<T> extends BinaryExpression<T, Number> {
   public Divide(Expression<T, Number> left, Expression<T, Number> right) {
