+++ src/com/jsheets/expressions/Expression.java
@@ -1,4 +1,4 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public interface Expression<T, R> {
   public R compute();
