+++ src/com/jsheets/expressions/NumericExpression.java
@@ -1,3 +1,3 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public interface NumericExpression extends Expression<Number, Number> {}
