+++ src/com/jsheets/expressions/Parentheses.java
@@ -1,4 +1,4 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public class Parentheses<T, R> extends UnaryExpression<T, R> {
   public Parentheses(Expression<T, R> expression) {
