+++ src/com/jsheets/expressions/Subtract.java
@@ -1,4 +1,4 @@
-package com.jsheets.components.expressions;
+package com.jsheets.expressions;
 
 public class Subtract<T> extends BinaryExpression<T, Number> {
   public Subtract(Expression<T, Number> left, Expression<T, Number> right) {
